Update Order Components

+++ src/main/java/com/example/oopone/service/OrderService.java
@@ -1,14 +1,22 @@
 package com.example.oopone.service;
 
+import com.example.oopone.dto.OrderDto;
+import com.example.oopone.dto.OrderItemsDto;
+import com.example.oopone.dto.UserDto;
 import com.example.oopone.model.CartItems;
 import com.example.oopone.model.OrderItems;
 import com.example.oopone.model.Orders;
+import com.example.oopone.model.User;
 import com.example.oopone.repository.CartItemsRepo;
 import com.example.oopone.repository.OrderItemsRepo;
 import com.example.oopone.repository.OrdersRepo;
+import com.example.oopone.repository.UserRepo;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import com.example.oopone.exception.SumExceedsException;
+import com.example.oopone.exception.CartEmptyException;
+import com.example.oopone.exception.QuantityExceedsException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -27,12 +35,42 @@ public class OrderService {
     @Autowired
     OrdersRepo ordersRepo;
 
-    public void addOrderedProducts(int user_id) {
+    @Autowired
+    UserRepo userRepo;
+
+    @Autowired
+    CartItemsService cartItemsService;
+
+    public String addOrderedProducts(int user_id) {
         List<CartItems> cartItemsList = cartItemsRepo.findAllByUserid(user_id);
-        Orders orders = new Orders(user_id);
-        for(CartItems var: cartItemsList){
-            OrderItems orderItems = new OrderItems(
-                    var.getQuantity_in_cart(),orders,var.getItem()
+        if(cartItemsList.isEmpty()){
+            throw new CartEmptyException("The Cart is Empty");
+        }
+        else {
+            User user = userRepo.findById(user_id).get();
+            double wallet_amount = user.getWallet_amt();
+            double sum = cartItemsService.calculateCartValue(user_id);
+            if (sum > wallet_amount) {
+                throw new SumExceedsException("The Sum Exceeds the Amount you have in your wallet");
+            } else {
+
+                for(CartItems var : cartItemsList){
+
+                    if(var.getQuantity_in_cart() > var.getItem().getQty_avlb()) {
+                        throw new QuantityExceedsException("Quantity Ordered for item: " + var.getItem().getId() +" more than Quantity Available");
+                    }
+                }
+                for(CartItems var : cartItemsList){
+                    var.getItem().setQty_avlb(var.getItem().getQty_avlb() - var.getQuantity_in_cart());
+
+                }
+
+
+
+                Orders orders = new Orders(user_id);
+                for (CartItems var : cartItemsList) {
+                    OrderItems orderItems = new OrderItems(
+                            var.getQuantity_in_cart(), orders, var.getItem()
 
                     );
                     cartItemsRepo.deleteById(var.getId());
